feat: simplify Supabase architecture for cleaner data flow
- Centralize all TypeScript types in src/lib/types.ts
- Extract recipe parsing logic to dedicated module (src/lib/recipe-parser.ts)
- Simplify Supabase client configuration with basic validation
- Streamline API layer with consistent error handling
- Update auth utilities to use centralized types
- Fix all import paths to use new type locations
- Create simplified database schema migration
- Add comprehensive architecture documentation
- Ensure all components use centralized types
- Maintain backward compatibility with re-exports
- Fix TypeScript linting issues with proper type annotations
- Apply Prettier formatting to all files

This creates a clean, straightforward connection between frontend and backend
while preserving all essential functionality.
