+++ src/lib/api.ts
@@ -1,5 +1,13 @@
 import { supabase } from './supabase';
-import type { Recipe, PublicRecipe } from './supabase';
+import type { Recipe, PublicRecipe } from './types';
+import { parseRecipeFromText } from './recipe-parser';
+
+// Simple error handler
+function handleError(error: unknown, operation: string): never {
+  console.error(`${operation} error:`, error);
+  const errorMessage = error instanceof Error ? error.message : 'Unknown error';
+  throw new Error(`${operation} failed: ${errorMessage}`);
+}
 
 export const recipeApi = {
   // Fetch all recipes for the current user
@@ -15,7 +23,7 @@ export const recipeApi = {
       .eq('user_id', user.id)
       .order('created_at', { ascending: false });
 
-    if (error) throw error;
+    if (error) handleError(error, 'Get user recipes');
     return data || [];
   },
 
@@ -27,20 +35,20 @@ export const recipeApi = {
       .eq('id', id)
       .single();
 
-    if (error) throw error;
+    if (error) handleError(error, 'Get recipe');
     return data;
   },
 
   // Fetch public recipes for the Explore feed
   async getPublicRecipes(): Promise<PublicRecipe[]> {
-    // First, get all public recipes
+    // Get all public recipes
     const { data: recipes, error: recipesError } = await supabase
       .from('recipes')
       .select('*')
       .eq('is_public', true)
       .order('created_at', { ascending: false });
 
-    if (recipesError) throw recipesError;
+    if (recipesError) handleError(recipesError, 'Get public recipes');
     if (!recipes || recipes.length === 0) return [];
 
     // Get unique user IDs from recipes
@@ -52,7 +60,7 @@ export const recipeApi = {
       .select('id, full_name')
       .in('id', userIds);
 
-    if (profilesError) throw profilesError;
+    if (profilesError) handleError(profilesError, 'Get profiles');
 
     // Create a map of user_id to full_name
     const profileMap = new Map(
@@ -60,12 +68,10 @@ export const recipeApi = {
     );
 
     // Combine recipes with profile data
-    const publicRecipes: PublicRecipe[] = recipes.map((recipe) => ({
+    return recipes.map((recipe) => ({
       ...recipe,
       author_name: profileMap.get(recipe.user_id) || 'Unknown Author',
     }));
-
-    return publicRecipes;
   },
 
   // Toggle recipe public status
@@ -75,7 +81,7 @@ export const recipeApi = {
       .update({ is_public: isPublic })
       .eq('id', recipeId);
 
-    if (error) throw error;
+    if (error) handleError(error, 'Toggle recipe public status');
   },
 
   // Get recipe sharing status
@@ -86,7 +92,7 @@ export const recipeApi = {
       .eq('id', recipeId)
       .single();
 
-    if (error) throw error;
+    if (error) handleError(error, 'Get recipe sharing status');
     return data?.is_public || false;
   },
 
@@ -105,7 +111,7 @@ export const recipeApi = {
       .select()
       .single();
 
-    if (error) throw error;
+    if (error) handleError(error, 'Create recipe');
     return data;
   },
 
@@ -118,7 +124,7 @@ export const recipeApi = {
       .select()
       .single();
 
-    if (error) throw error;
+    if (error) handleError(error, 'Update recipe');
     return data;
   },
 
@@ -126,7 +132,7 @@ export const recipeApi = {
   async deleteRecipe(id: string): Promise<void> {
     const { error } = await supabase.from('recipes').delete().eq('id', id);
 
-    if (error) throw error;
+    if (error) handleError(error, 'Delete recipe');
   },
 
   // Save (clone) a public recipe to user's collection
@@ -136,7 +142,7 @@ export const recipeApi = {
     } = await supabase.auth.getUser();
     if (!user) throw new Error('User not authenticated');
 
-    // First, get the public recipe
+    // Get the public recipe
     const { data: sourceRecipe, error: fetchError } = await supabase
       .from('recipes')
       .select('*')
@@ -144,7 +150,7 @@ export const recipeApi = {
       .eq('is_public', true)
       .single();
 
-    if (fetchError) throw fetchError;
+    if (fetchError) handleError(fetchError, 'Fetch public recipe');
     if (!sourceRecipe) throw new Error('Recipe not found or not public');
 
     // Create a new recipe with the current user as owner
@@ -153,9 +159,9 @@ export const recipeApi = {
       ingredients: sourceRecipe.ingredients,
       instructions: sourceRecipe.instructions,
       notes: sourceRecipe.notes,
-      image_url: sourceRecipe.image_url, // Keep the same image URL for now
+      image_url: sourceRecipe.image_url,
       user_id: user.id,
-      is_public: false, // Make it private by default
+      is_public: false,
     };
 
     const { data, error } = await supabase
@@ -164,7 +170,7 @@ export const recipeApi = {
       .select()
       .single();
 
-    if (error) throw error;
+    if (error) handleError(error, 'Save public recipe');
     return data;
   },
 
@@ -182,7 +188,7 @@ export const recipeApi = {
       .from('recipe-images')
       .upload(fileName, file);
 
-    if (uploadError) throw uploadError;
+    if (uploadError) handleError(uploadError, 'Upload image');
 
     const { data } = supabase.storage
       .from('recipe-images')
@@ -190,384 +196,7 @@ export const recipeApi = {
 
     return data.publicUrl;
   },
+
+  // Parse recipe from text (delegates to recipe-parser)
+  parseRecipeFromText,
 };
-
-// Types for the new JSON recipe format
-interface IngredientItem {
-  item: string;
-  amount?: string;
-  prep?: string;
-}
-
-// RecipeJSON interface removed as it's not used directly due to dynamic JSON parsing
-
-// Parse recipe from text using external API
-export async function parseRecipeFromText(text: string): Promise<{
-  title: string;
-  ingredients: string[];
-  instructions: string;
-  notes: string;
-}> {
-  // This would typically call your LLM proxy endpoint
-  // For now, we'll simulate the parsing with a basic implementation
-  try {
-    console.log('Attempting to parse as JSON...');
-
-    let jsonText = text;
-
-    // Check if JSON is wrapped in markdown code blocks
-    const jsonBlockMatch = text.match(/```json\s*([\s\S]*?)\s*```/);
-    if (jsonBlockMatch) {
-      jsonText = jsonBlockMatch[1];
-      console.log('Extracted JSON from markdown code block');
-    }
-
-    // Try to parse as JSON first
-    const parsed = JSON.parse(jsonText);
-
-    console.log('Successfully parsed JSON format');
-
-    // Validate required fields - be flexible about field names
-    const hasTitle = parsed.title || parsed.name;
-    const hasIngredients =
-      parsed.ingredients && Array.isArray(parsed.ingredients);
-    const hasInstructions = parsed.instructions;
-
-    if (!hasTitle) {
-      throw new Error('Missing required field: title or name');
-    }
-    if (!hasIngredients) {
-      throw new Error('Missing required field: ingredients (must be an array)');
-    }
-    if (!hasInstructions) {
-      throw new Error('Missing required field: instructions');
-    }
-
-    // Handle complex nested JSON structure
-    if (parsed.name || parsed.title) {
-      const title = parsed.name || parsed.title || 'Untitled Recipe';
-      let ingredients: string[] = [];
-      let instructions = '';
-      let notes = '';
-
-      // Handle nested ingredients object with categories
-      if (parsed.ingredients && typeof parsed.ingredients === 'object') {
-        if (Array.isArray(parsed.ingredients)) {
-          // Simple array format - convert to strings
-          ingredients = parsed.ingredients.map(
-            (item: string | IngredientItem) =>
-              typeof item === 'string'
-                ? item
-                : `${item.amount || ''} ${item.item || ''} ${item.prep ? `, ${item.prep}` : ''}`.trim()
-          );
-        } else {
-          // Nested object format with categories (main, sauce, toppings, etc.)
-          const categoryOrder = ['main', 'sauce', 'toppings', 'garnish']; // Preferred order
-          const allCategories = Object.keys(parsed.ingredients);
-
-          // Process categories in preferred order, then any remaining
-          const orderedCategories = [
-            ...categoryOrder.filter((cat) => allCategories.includes(cat)),
-            ...allCategories.filter((cat) => !categoryOrder.includes(cat)),
-          ];
-
-          for (const category of orderedCategories) {
-            const items = parsed.ingredients[category];
-            if (Array.isArray(items) && items.length > 0) {
-              // Add category header with proper capitalization
-              const categoryTitle = category
-                .replace(/([A-Z])/g, ' $1')
-                .replace(/^./, (str) => str.toUpperCase());
-              ingredients.push(`--- ${categoryTitle} ---`);
-
-              for (const item of items) {
-                if (typeof item === 'string') {
-                  ingredients.push(item);
-                } else if (typeof item === 'object' && item !== null) {
-                  // Handle structured ingredient objects: { item, amount, prep }
-                  const typedItem = item as IngredientItem;
-                  let ingredientStr = '';
-                  if (typedItem.amount) ingredientStr += `${typedItem.amount} `;
-                  if (typedItem.item) ingredientStr += typedItem.item;
-                  if (typedItem.prep) ingredientStr += `, ${typedItem.prep}`;
-
-                  if (ingredientStr.trim()) {
-                    ingredients.push(ingredientStr.trim());
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-
-      // Handle instructions - support both array and string formats
-      const instructionParts: string[] = [];
-
-      // Add basic/preparation instructions first (if present)
-      if (
-        parsed.basic_instructions &&
-        Array.isArray(parsed.basic_instructions) &&
-        parsed.basic_instructions.length > 0
-      ) {
-        instructionParts.push('**Preparation:**');
-        parsed.basic_instructions.forEach((step: string, index: number) => {
-          if (step && step.trim()) {
-            instructionParts.push(`${index + 1}. ${step.trim()}`);
-          }
-        });
-        instructionParts.push(''); // Add blank line
-      }
-
-      // Add main cooking instructions
-      if (
-        Array.isArray(parsed.instructions) &&
-        parsed.instructions.length > 0
-      ) {
-        if (instructionParts.length > 0) {
-          instructionParts.push('**Cooking Instructions:**');
-        }
-        parsed.instructions.forEach((step: string) => {
-          if (step && step.trim()) {
-            // Remove any existing numbering and add clean formatting
-            const cleanStep = step.trim().replace(/^\d+\.\s*/, '');
-            instructionParts.push(cleanStep);
-          }
-        });
-      } else if (
-        typeof parsed.instructions === 'string' &&
-        parsed.instructions.trim()
-      ) {
-        if (instructionParts.length > 0) {
-          instructionParts.push('**Cooking Instructions:**');
-        }
-        instructionParts.push(parsed.instructions.trim());
-      }
-
-      instructions = instructionParts.join('\n');
-
-      // Handle notes - combine various optional sections
-      const notesParts: string[] = [];
-
-      // Add servings information
-      if (parsed.servings) {
-        notesParts.push(`**Servings:** ${parsed.servings}`);
-        notesParts.push('');
-      }
-
-      // Add tips and tricks
-      if (
-        parsed.tips_and_tricks &&
-        Array.isArray(parsed.tips_and_tricks) &&
-        parsed.tips_and_tricks.length > 0
-      ) {
-        notesParts.push('**Tips & Tricks:**');
-        parsed.tips_and_tricks.forEach((tip: string) => {
-          if (tip && tip.trim()) {
-            notesParts.push(`• ${tip.trim()}`);
-          }
-        });
-        notesParts.push('');
-      }
-
-      // Add substitutions
-      if (
-        parsed.substitutions &&
-        Array.isArray(parsed.substitutions) &&
-        parsed.substitutions.length > 0
-      ) {
-        notesParts.push('**Substitutions:**');
-        parsed.substitutions.forEach((sub: string) => {
-          if (sub && sub.trim()) {
-            notesParts.push(`• ${sub.trim()}`);
-          }
-        });
-        notesParts.push('');
-      }
-
-      // Add pairings
-      if (
-        parsed.pairings &&
-        Array.isArray(parsed.pairings) &&
-        parsed.pairings.length > 0
-      ) {
-        notesParts.push('**Pairings:**');
-        parsed.pairings.forEach((pairing: string) => {
-          if (pairing && pairing.trim()) {
-            notesParts.push(`• ${pairing.trim()}`);
-          }
-        });
-        notesParts.push('');
-      }
-
-      // Add any additional notes
-      if (
-        parsed.notes &&
-        typeof parsed.notes === 'string' &&
-        parsed.notes.trim()
-      ) {
-        notesParts.push('**Additional Notes:**');
-        notesParts.push(parsed.notes.trim());
-      }
-
-      notes = notesParts.join('\n').trim();
-
-      return {
-        title,
-        ingredients,
-        instructions,
-        notes,
-      };
-    }
-
-    // Fallback for simple JSON format
-    return {
-      title: parsed.title || parsed.name || 'Untitled Recipe',
-      ingredients: Array.isArray(parsed.ingredients) ? parsed.ingredients : [],
-      instructions: parsed.instructions || '',
-      notes: parsed.notes || '',
-    };
-  } catch {
-    console.log('JSON parsing failed, trying markdown parsing');
-    console.log('Text preview:', text.substring(0, 200) + '...');
-
-    // Try to extract recipe content from the text if it contains non-recipe content
-    let processedText = text;
-
-    // Look for recipe-like sections and extract them
-    const recipeKeywords = [
-      'ingredient',
-      'instruction',
-      'recipe',
-      'cook',
-      'bake',
-      'prepare',
-      'serve',
-      'mix',
-      'add',
-      'heat',
-      'boil',
-      'fry',
-      'season',
-      'cup',
-      'tablespoon',
-      'teaspoon',
-      'oz',
-      'lb',
-      'gram',
-    ];
-    const lines = text.split('\n');
-
-    // Find the first line that contains recipe-related content
-    let recipeStartIndex = -1;
-    for (let i = 0; i < lines.length; i++) {
-      const line = lines[i].toLowerCase();
-      if (recipeKeywords.some((keyword) => line.includes(keyword))) {
-        recipeStartIndex = i;
-        break;
-      }
-    }
-
-    // If we found recipe content, start from there
-    if (recipeStartIndex >= 0) {
-      processedText = lines.slice(recipeStartIndex).join('\n');
-      console.log(
-        'Extracted recipe content starting from line',
-        recipeStartIndex
-      );
-    } else {
-      // If no clear recipe content found, try to extract any structured content
-      console.log('No clear recipe keywords found, using full text');
-    }
-
-    // If not JSON, treat as markdown and do basic parsing
-    const processedLines = processedText
-      .split('\n')
-      .filter((line) => line.trim());
-    let title = 'Untitled Recipe';
-    const ingredients: string[] = [];
-    let instructions = '';
-    let notes = '';
-
-    let currentSection = '';
-
-    for (const line of processedLines) {
-      const trimmed = line.trim();
-      const lowerTrimmed = trimmed.toLowerCase();
-
-      if (trimmed.startsWith('#')) {
-        // Only use as title if it looks like a recipe title
-        const potentialTitle = trimmed.replace(/^#+\s*/, '');
-        if (
-          potentialTitle.length > 0 &&
-          !lowerTrimmed.includes('implementation') &&
-          !lowerTrimmed.includes('outcome')
-        ) {
-          title = potentialTitle;
-          currentSection = 'title';
-        }
-      } else if (lowerTrimmed.includes('ingredient')) {
-        currentSection = 'ingredients';
-      } else if (
-        lowerTrimmed.includes('instruction') ||
-        lowerTrimmed.includes('direction') ||
-        lowerTrimmed.includes('step') ||
-        lowerTrimmed.includes('method')
-      ) {
-        currentSection = 'instructions';
-      } else if (
-        lowerTrimmed.includes('note') ||
-        lowerTrimmed.includes('tip')
-      ) {
-        currentSection = 'notes';
-      } else if (
-        trimmed.startsWith('-') ||
-        trimmed.startsWith('*') ||
-        /^\d+\./.test(trimmed)
-      ) {
-        if (currentSection === 'ingredients') {
-          ingredients.push(trimmed.replace(/^[-*]\s*|\d+\.\s*/, ''));
-        } else if (currentSection === 'instructions') {
-          instructions += trimmed.replace(/^[-*]\s*|\d+\.\s*/, '') + '\n';
-        }
-      } else if (trimmed) {
-        if (currentSection === 'instructions') {
-          instructions += trimmed + '\n';
-        } else if (currentSection === 'notes') {
-          notes += trimmed + '\n';
-        } else if (
-          !currentSection &&
-          title === 'Untitled Recipe' &&
-          !lowerTrimmed.includes('implementation') &&
-          !lowerTrimmed.includes('outcome') &&
-          !lowerTrimmed.includes('analysis') &&
-          trimmed.length < 100
-        ) {
-          // Only use as title if it's short and doesn't contain non-recipe words
-          title = trimmed;
-        }
-      }
-    }
-
-    // If we didn't find much recipe content, provide a helpful default
-    if (ingredients.length === 0 && instructions.trim() === '') {
-      return {
-        title: 'Recipe from AI Conversation',
-        ingredients: ['Please ask the AI to provide specific ingredients'],
-        instructions:
-          'Please ask the AI to provide step-by-step cooking instructions.',
-        notes: `Original AI response: ${text.substring(0, 500)}${text.length > 500 ? '...' : ''}\n\nTip: Try asking the AI something like "Can you give me a complete recipe with ingredients and instructions?"`,
-      };
-    }
-
-    const result = {
-      title,
-      ingredients,
-      instructions: instructions.trim(),
-      notes: notes.trim(),
-    };
-
-    console.log('Markdown parsing result:', result);
-    return result;
-  }
-}
