+++ src/lib/supabase.ts
@@ -1,46 +1,25 @@
 import { createClient } from '@supabase/supabase-js';
+import type {
+  Recipe,
+  PublicRecipe,
+  Profile,
+  Username,
+  UserSafety,
+  CookingPreferences,
+  AccountEvent,
+} from './types';
 
+// Environment variable validation
 const supabaseUrl = import.meta.env.VITE_SUPABASE_URL;
 const supabaseAnonKey = import.meta.env.VITE_SUPABASE_ANON_KEY;
 
-// Better error handling for missing environment variables
 if (!supabaseUrl || !supabaseAnonKey) {
-  console.error('❌ Missing Supabase environment variables:');
-  console.error('   VITE_SUPABASE_URL:', supabaseUrl ? 'SET' : 'MISSING');
-  console.error(
-    '   VITE_SUPABASE_ANON_KEY:',
-    supabaseAnonKey ? 'SET' : 'MISSING'
-  );
-
-  if (
-    supabaseUrl &&
-    (supabaseUrl.includes('127.0.0.1') || supabaseUrl.includes('localhost'))
-  ) {
-    console.error('❌ ERROR: Supabase URL points to localhost!');
-    console.error('   This should point to your production Supabase project.');
-    console.error('   Safari blocks localhost connections from HTTPS sites.');
-    console.error('   Please update your Vercel environment variables.');
-  }
-
   throw new Error(
-    'Missing Supabase environment variables. Please check your Vercel configuration.'
+    'Missing Supabase environment variables. Please check your configuration.'
   );
 }
 
-// Validate Supabase URL format
-if (
-  !supabaseUrl.startsWith('https://') &&
-  !supabaseUrl.startsWith('http://127.0.0.1')
-) {
-  console.error(
-    '❌ Invalid Supabase URL format. Should start with https:// or http://127.0.0.1 for local development'
-  );
-  console.error('   Safari requires HTTPS for production sites.');
-  throw new Error(
-    'Invalid Supabase URL format - must use HTTPS for production or http://127.0.0.1 for local development'
-  );
-}
-
+// Create Supabase client with simplified configuration
 export const supabase = createClient(supabaseUrl, supabaseAnonKey, {
   auth: {
     persistSession: true,
@@ -49,80 +28,13 @@ export const supabase = createClient(supabaseUrl, supabaseAnonKey, {
   },
 });
 
-// Type definitions
-export type Recipe = {
-  id: string;
-  title: string;
-  ingredients: string[];
-  instructions: string;
-  notes: string;
-  image_url: string | null;
-  user_id: string;
-  is_public: boolean;
-  created_at: string;
-  updated_at: string;
-};
-
-export type PublicRecipe = Recipe & {
-  author_name: string;
-};
-
-export type Profile = {
-  id: string;
-  username: string | null;
-  full_name: string | null;
-  avatar_url: string | null;
-  bio: string | null;
-  // Phase 1A: Basic preferences
-  region: string | null;
-  language: string | null;
-  units: string | null;
-  time_per_meal: number | null;
-  skill_level: string | null;
-  created_at: string;
-  updated_at: string;
-};
-
-export type Username = {
-  username: string;
-  user_id: string;
-  created_at: string;
-};
-
-// Phase 1B: Safety data type
-export type UserSafety = {
-  user_id: string;
-  allergies: string[];
-  dietary_restrictions: string[];
-  medical_conditions: string[];
-  created_at: string;
-  updated_at: string;
-};
-
-// Phase 1C: Cooking preferences type
-export type CookingPreferences = {
-  user_id: string;
-  preferred_cuisines: string[];
-  available_equipment: string[];
-  disliked_ingredients: string[];
-  spice_tolerance: number | null;
-  created_at: string;
-  updated_at: string;
-};
-
-export type AccountEvent = {
-  id: string;
-  user_id: string;
-  event_type:
-    | 'profile_created'
-    | 'username_claimed'
-    | 'username_changed'
-    | 'profile_updated'
-    | 'avatar_updated'
-    | 'email_changed'
-    | 'password_changed';
-  metadata: Record<string, unknown>;
-  ip_address: string | null;
-  user_agent: string | null;
-  created_at: string;
+// Re-export types for backward compatibility
+export type {
+  Recipe,
+  PublicRecipe,
+  Profile,
+  Username,
+  UserSafety,
+  CookingPreferences,
+  AccountEvent,
 };
